update main.js so user sets params for auto
Updated main.js so that the user can set the parameters for automated trading. User is now prompted to select a pair to trade on, order size for the pair, the current price of asset (bot can fetch, but not from w/i the function as needed but needed to calculate the point to take profit and the point to buy and cost average down), % takeProfit, and % costAverageDown.